role changes added

+++ job.portal/src/main/java/job/portal/entities/Role.java
@@ -14,7 +14,7 @@ public class Role {
 
     @Id
     @Column(name = "role_id")
-    private int roleId;
+    private Long roleId;
 
     @Column(name = "role_name", nullable = false, unique = true, length = 50)
     private String roleName;
